split EpisodeEventWrapper into client and server parts
