+++ src/main/java/io/singularitynet/EpisodeEventWrapper.java
@@ -36,20 +36,17 @@ import java.util.concurrent.locks.ReentrantReadWriteLock;
  * and passing them on to the current episode.<br>
  * Doing it this way saves us having to register/deregister each individual episode, which was causing race-condition vulnerabilities.
  */
-public class EpisodeEventWrapper implements ClientTickEvents.EndTick,
+public class EpisodeEventWrapper implements
         ServerTickEvents.EndTick,
-        ClientChunkEvents.Load,
-        ClientLifecycleEvents.ClientStarted,
-        TitleScreenEvents.EndInit,
         ServerTickEvents.StartTick {
     /** The current episode, if there is one. */
-    private StateEpisode stateEpisode = null;
+    protected StateEpisode stateEpisode = null;
 
     /** Lock to prevent the state episode being changed whilst mid event.<br>
      * This does not prevent multiple events from *reading* the stateEpisode, but
      * it won't allow *writing* to the stateEpisode whilst it is being read.
      */
-    ReentrantReadWriteLock stateEpisodeLock = new ReentrantReadWriteLock();
+    protected ReentrantReadWriteLock stateEpisodeLock = new ReentrantReadWriteLock();
 
     /** Set our state to a new episode.<br>
      * This waits on the stateEpisodeLock to prevent the episode being changed whilst in use.
@@ -64,72 +61,11 @@ public class EpisodeEventWrapper implements ClientTickEvents.EndTick,
         this.stateEpisodeLock.writeLock().unlock();
         return lastEpisode;
     }
-/*
-    @SubscribeEvent
-    public void onPlayerTick(TickEvent.PlayerTickEvent ev)
-    {
-        // Pass the event on to the active episode, if there is one:
-        this.stateEpisodeLock.readLock().lock();
-        if (this.stateEpisode != null && this.stateEpisode.isLive())
-        {
-            this.stateEpisode.onPlayerTick(ev);
-        }
-        this.stateEpisodeLock.readLock().unlock();
-    }*/
 
-
-    public void onRenderTickEnd(WorldRenderContext ev)
-    {
-        // Pass the event on to the active episode, if there is one:
-        this.stateEpisodeLock.readLock().lock();
-        if (this.stateEpisode != null && this.stateEpisode.isLive())
-        {
-            this.stateEpisode.onRenderTickEnd(ev);
-        }
-        this.stateEpisodeLock.readLock().unlock();
-    }
-
-    public void onRenderTickStart(WorldRenderContext ev)
-    {
-        // Pass the event on to the active episode, if there is one:
-        this.stateEpisodeLock.readLock().lock();
-        if (this.stateEpisode != null && this.stateEpisode.isLive())
-        {
-            this.stateEpisode.onRenderTickStart(ev);
-        }
-        this.stateEpisodeLock.readLock().unlock();
-    }
-
-    // use https://github.com/ByMartrixx/player-events
-    /*
-    @SubscribeEvent
-    public void onPlayerDies(LivingDeathEvent lde)
-    {
-        // Pass the event on to the active episode, if there is one:
-        this.stateEpisodeLock.readLock().lock();
-        if (this.stateEpisode != null && this.stateEpisode.isLive())
-        {
-            this.stateEpisode.onPlayerDies(lde);
-        }
-        this.stateEpisodeLock.readLock().unlock();
-    }*/
-
-    @Override
-    public void onEndTick(MinecraftClient client) {
-        // Now pass the event on to the active episode, if there is one:
-        this.stateEpisodeLock.readLock().lock();
-        if (this.stateEpisode != null && this.stateEpisode.isLive())
-        {
-            try
-            {
-                this.stateEpisode.onClientTick(client);
-            }
-            catch (Exception e)
-            {
-                LogManager.getLogger().error("Exception onClientTick", e);
-            }
-        }
-        this.stateEpisodeLock.readLock().unlock();
+    public EpisodeEventWrapper(){
+        LogManager.getLogger().info("Setting up EpisodeEventWrapper for Server events");
+        ServerTickEvents.END_SERVER_TICK.register(server -> {this.onEndTick(server);});
+        ServerTickEvents.START_SERVER_TICK.register(server -> {this.onStartTick(server);});
     }
 
     @Override
@@ -139,39 +75,6 @@ public class EpisodeEventWrapper implements ClientTickEvents.EndTick,
         if (this.stateEpisode != null && this.stateEpisode.isLive())
         {
             this.stateEpisode.onServerTick(server);
-        }
-        this.stateEpisodeLock.readLock().unlock();
-    }
-
-    @Override
-    public void onChunkLoad(ClientWorld world, WorldChunk chunk) {
-        // Pass the event on to the active episode, if there is one:
-        this.stateEpisodeLock.readLock().lock();
-        if (this.stateEpisode != null && this.stateEpisode.isLive())
-        {
-            this.stateEpisode.onChunkLoad(world, chunk);
-        }
-        this.stateEpisodeLock.readLock().unlock();
-    }
-
-    @Override
-    public void onClientStarted(MinecraftClient client) {
-        // Pass the event on to the active episode, if there is one:
-        this.stateEpisodeLock.readLock().lock();
-        if (this.stateEpisode != null && this.stateEpisode.isLive())
-        {
-            this.stateEpisode.onClientStarted(client);
-        }
-        this.stateEpisodeLock.readLock().unlock();
-    }
-
-    @Override
-    public void onTitleScreenEndInit(){
-        // Pass the event on to the active episode, if there is one:
-        this.stateEpisodeLock.readLock().lock();
-        if (this.stateEpisode != null && this.stateEpisode.isLive())
-        {
-            this.stateEpisode.onTitleScreen();
         }
         this.stateEpisodeLock.readLock().unlock();
     }
